+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- Changed app name from filearbgen to flutter_localization_generator
+- Added splitElement add feature
+- Fixed Error texts